Removed class LogicalDeviceType

+++ src/MyPeer.cpp
@@ -371,7 +371,7 @@ bool MyPeer::load(BaseLib::Systems::ICentral* central)
 		loadVariables(central, rows);
 		if(!_rpcDevice)
 		{
-			GD::out.printError("Error loading peer " + std::to_string(_peerID) + ": Device type not found: 0x" + BaseLib::HelperFunctions::getHexString((uint32_t)_deviceType.type()) + " Firmware version: " + std::to_string(_firmwareVersion));
+			GD::out.printError("Error loading peer " + std::to_string(_peerID) + ": Device type not found: 0x" + BaseLib::HelperFunctions::getHexString(_deviceType) + " Firmware version: " + std::to_string(_firmwareVersion));
 			return false;
 		}
